+++ src/db/list_ng_result.cc
@@ -11,7 +11,7 @@ namespace db {
 	void ListNGResult::printToConnection(Connection &conn) throw (ConnectionClosedException){
 		conn.write(Protocol::ANS_LIST_NG);
 		print_int(conn, news_groups.size());
-		for (int i = 0; i < news_groups.size(); ++i) {
+		for (size_t i = 0; i < news_groups.size(); ++i) {
 			pair<int, string> ng = news_groups[i];
 			print_int(conn, ng.first);
 			print_string(conn, ng.second);
